Tracker refactored
